FAC-735: refactor - Resolve the conflicts after merging
