+++ flickit-assessment-kit/src/test/java/org/flickit/assessment/kit/application/service/answerrange/UpdateAnswerRangeServiceTest.java
@@ -1,11 +1,11 @@
 package org.flickit.assessment.kit.application.service.answerrange;
 
 import org.flickit.assessment.common.exception.AccessDeniedException;
-import org.flickit.assessment.common.exception.ResourceNotFoundException;
 import org.flickit.assessment.common.exception.ValidationException;
+import org.flickit.assessment.kit.application.domain.AnswerRange;
 import org.flickit.assessment.kit.application.domain.KitVersion;
 import org.flickit.assessment.kit.application.port.in.answerrange.UpdateAnswerRangeUseCase;
-import org.flickit.assessment.kit.application.port.out.answerange.LoadAnswerRangePort;
+import org.flickit.assessment.kit.application.port.out.answerrange.LoadAnswerRangePort;
 import org.flickit.assessment.kit.application.port.out.answerrange.UpdateAnswerRangePort;
 import org.flickit.assessment.kit.application.port.out.expertgroup.LoadExpertGroupOwnerPort;
 import org.flickit.assessment.kit.application.port.out.kitversion.LoadKitVersionPort;
@@ -18,7 +18,6 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import java.util.Optional;
 import java.util.UUID;
 import java.util.function.Consumer;
 
@@ -52,6 +51,7 @@ class UpdateAnswerRangeServiceTest {
 
     private final UUID ownerId = UUID.randomUUID();
     private final KitVersion kitVersion = createKitVersion(simpleKit());
+    private final AnswerRange answerRange = AnswerRangeMother.createReusableAnswerRange();
 
     @Test
     void testUpdateAnswerRange_WhenCurrentUserIsNotExpertGroupOwner_ThenThrowAccessDeniedException() {
@@ -69,11 +69,10 @@ class UpdateAnswerRangeServiceTest {
     @Test
     void testUpdateAnswerRange_WhenAnswerRangeIsUsedByQuestionsAndReusableToNonReusableUpdate_ThenThrowValidationException() {
         var param = createParam(b -> b.currentUserId(ownerId).reusable(false));
-        var answerRange = AnswerRangeMother.createAnswerRangeWithIsReusable(true);
 
         when(loadKitVersionPort.load(param.getKitVersionId())).thenReturn(kitVersion);
         when(loadExpertGroupOwnerPort.loadOwnerId(kitVersion.getKit().getExpertGroupId())).thenReturn(ownerId);
-        when(loadAnswerRangePort.loadAnswerRange(param.getAnswerRangeId(), param.getKitVersionId())).thenReturn(Optional.of(answerRange));
+        when(loadAnswerRangePort.load(param.getAnswerRangeId(), param.getKitVersionId())).thenReturn(answerRange);
         when(checkQuestionExistencePort.checkByAnswerRange(param.getAnswerRangeId())).thenReturn(true);
 
         var throwable = assertThrows(ValidationException.class, () -> service.updateAnswerRange(param));
@@ -85,11 +84,10 @@ class UpdateAnswerRangeServiceTest {
     @Test
     void testUpdateAnswerRange_WhenCurrentUserIsExpertGroupOwnerAndReusableAnswerRangeTitleIsNull_ThenThrowValidationException() {
         var param = createParam(b -> b.currentUserId(ownerId).title(null));
-        var answerRange = AnswerRangeMother.createAnswerRangeWithIsReusable(true);
 
         when(loadKitVersionPort.load(param.getKitVersionId())).thenReturn(kitVersion);
         when(checkQuestionExistencePort.checkByAnswerRange(param.getAnswerRangeId())).thenReturn(false);
-        when(loadAnswerRangePort.loadAnswerRange(param.getAnswerRangeId(), param.getKitVersionId())).thenReturn(Optional.of(answerRange));
+        when(loadAnswerRangePort.load(param.getAnswerRangeId(), param.getKitVersionId())).thenReturn(answerRange);
         when(loadExpertGroupOwnerPort.loadOwnerId(kitVersion.getKit().getExpertGroupId())).thenReturn(ownerId);
 
         var throwable = assertThrows(ValidationException.class, () -> service.updateAnswerRange(param));
@@ -99,27 +97,12 @@ class UpdateAnswerRangeServiceTest {
     }
 
     @Test
-    void testUpdateAnswerRange_WhenAnswerRangeNotFound_ResourceNotFoundException() {
-        var param = createParam(b -> b.currentUserId(ownerId).title(null));
-
-        when(loadKitVersionPort.load(param.getKitVersionId())).thenReturn(kitVersion);
-        when(loadExpertGroupOwnerPort.loadOwnerId(kitVersion.getKit().getExpertGroupId())).thenReturn(ownerId);
-        when(loadAnswerRangePort.loadAnswerRange(param.getAnswerRangeId(), param.getKitVersionId())).thenReturn(Optional.empty());
-
-        var throwable = assertThrows(ResourceNotFoundException.class, () -> service.updateAnswerRange(param));
-        assertEquals(UPDATE_ANSWER_RANGE_ANSWER_RANGE_ID_NOT_FOUND, throwable.getMessage());
-
-        verifyNoInteractions(updateAnswerRangePort);
-    }
-
-    @Test
     void testUpdateAnswerRange_WhenCurrentUserIsExpertGroupOwnerAndReusableAnswerRangeTitleIsValid_ThenUpdateAnswerRange() {
         var param = createParam(b -> b.currentUserId(ownerId));
-        var answerRange = AnswerRangeMother.createAnswerRangeWithIsReusable(true);
 
         when(loadKitVersionPort.load(param.getKitVersionId())).thenReturn(kitVersion);
         when(loadExpertGroupOwnerPort.loadOwnerId(kitVersion.getKit().getExpertGroupId())).thenReturn(ownerId);
-        when(loadAnswerRangePort.loadAnswerRange(param.getAnswerRangeId(), param.getKitVersionId())).thenReturn(Optional.of(answerRange));
+        when(loadAnswerRangePort.load(param.getAnswerRangeId(), param.getKitVersionId())).thenReturn(answerRange);
         doNothing().when(updateAnswerRangePort).update(any(UpdateAnswerRangePort.Param.class));
 
         service.updateAnswerRange(param);
